Create scripts workspace so we can lint/type-check scripts
