+++ workspaces/scripts/.nvmrc
@@ -1,0 +1,1 @@
+22.14.0
