+++ workspaces/scripts/.prettierignore
@@ -1,0 +1,24 @@
+/node_modules/*
+/coverage/*
+/build/*
+/dist/*
+/.vscode/*
+*.txt
+*.snap
+*.toml
+*.png
+*.svg
+*.jpg
+*.jpeg
+*.gif
+
+.DS_Store
+.eslintignore
+.prettierignore
+.gitignore
+.gitattributes
+.npmignore
+.nvmrc
+.browserslistrc
+.env
+LICENSE.txt
