+++ workspaces/scripts/README.md
@@ -1,0 +1,5 @@
+# RCX Scripts
+
+**Scripts for one or more RCX workspaces**
+
+See contents of `src` directory for available scripts.
