+++ workspaces/scripts/eslint.config.js
@@ -1,0 +1,3 @@
+import eslintConfigRCXDev from '@blinkorb/eslint-config-rcx-dev';
+
+export default eslintConfigRCXDev;
